+++ Tasklist/app/src/main/java/es/ucm/fdi/tasklist/ui/today/TodayFragment.java
@@ -139,7 +139,7 @@ public class TodayFragment extends Fragment implements ObserverDao {
         }
     }
 
-    public TaskDetail updateList(boolean remove, int _id,  String _title, String _desc, String _date, boolean _fin, boolean _imp, String _hora){
+    public TaskDetail updateList(boolean remove, long _id,  String _title, String _desc, String _date, boolean _fin, boolean _imp, String _hora){
         TaskDetail detail = new TaskDetail(_id, _title, _desc, _date, _fin, _imp, _hora);
         if(remove) todayTaskList.remove(detail);
         else{
@@ -187,7 +187,7 @@ public class TodayFragment extends Fragment implements ObserverDao {
         String title, content, date, hora;
         boolean finish;
         boolean important;
-        int id;
+        long id;
 
         if(data != null) {
             title = data.getExtras().getString("title");
@@ -198,7 +198,7 @@ public class TodayFragment extends Fragment implements ObserverDao {
             important = data.getExtras().getBoolean("important");
 
             if (requestCode == 2) {
-                id = data.getExtras().getInt("id");
+                id = data.getExtras().getLong("id");
                 if (resultCode == Activity.RESULT_OK) {
                     TaskDetail td = updateList(false, id, title, content, date, finish, important, hora);
                     DataBaseTask.getInstance(getContext()).updateItem(td, db);
